fix delivery detail project size
Fixes #74

Abstracts out code to create a project summary into
ProjectSummaryFetcher. Uses this for both delivery-detail and
project-detail.
